Struct diem_sdk::client::BlockingClient

pub struct BlockingClient { /* private fields */ }

Implementations

impl BlockingClient

pub fn new<T>(url: T) -> BlockingClientwhere
T: Into<String>,

pub fn last_known_state(&self) -> Option<State>

pub fn wait_for_signed_transaction(
&self,
txn: &SignedTransaction,
timeout: Option<Duration>,
delay: Option<Duration>
) -> Result<Response<TransactionView>, WaitForTransactionError>

pub fn wait_for_transaction(
&self,
address: AccountAddress,
seq: u64,
expiration_time_secs: u64,
txn_hash: HashValue,
timeout: Option<Duration>,
delay: Option<Duration>
) -> Result<Response<TransactionView>, WaitForTransactionError>

pub fn batch(
&self,
requests: Vec<MethodRequest, Global>
) -> Result<Vec<Result<Response<MethodResponse>, Error>, Global>, Error>

pub fn request(
&self,
request: MethodRequest
) -> Result<Response<MethodResponse>, Error>

pub fn submit(&self, txn: &SignedTransaction) -> Result<Response<()>, Error>

pub fn get_metadata_by_version(
&self,
version: u64
) -> Result<Response<MetadataView>, Error>

pub fn get_metadata(&self) -> Result<Response<MetadataView>, Error>

pub fn get_account(
&self,
address: AccountAddress
) -> Result<Response<Option<AccountView>>, Error>

pub fn get_account_by_version(
&self,
address: AccountAddress,
version: u64
) -> Result<Response<Option<AccountView>>, Error>

pub fn get_transactions(
&self,
start_seq: u64,
limit: u64,
include_events: bool
) -> Result<Response<Vec<TransactionView, Global>>, Error>

pub fn get_account_transaction(
&self,
address: AccountAddress,
seq: u64,
include_events: bool
) -> Result<Response<Option<TransactionView>>, Error>

pub fn get_account_transactions(
&self,
address: AccountAddress,
start_seq: u64,
limit: u64,
include_events: bool
) -> Result<Response<Vec<TransactionView, Global>>, Error>

pub fn get_events(
&self,
key: EventKey,
start_seq: u64,
limit: u64
) -> Result<Response<Vec<EventView, Global>>, Error>

pub fn get_currencies(
&self
) -> Result<Response<Vec<CurrencyInfoView, Global>>, Error>

pub fn get_network_status(&self) -> Result<Response<u64>, Error>

pub fn get_state_proof(
&self,
from_version: u64
) -> Result<Response<StateProofView>, Error>

pub fn get_accumulator_consistency_proof(
&self,
client_known_version: Option<u64>,
ledger_version: Option<u64>
) -> Result<Response<AccumulatorConsistencyProofView>, Error>

pub fn get_account_state_with_proof(
&self,
address: AccountAddress,
from_version: Option<u64>,
to_version: Option<u64>
) -> Result<Response<AccountStateWithProofView>, Error>

pub fn get_transactions_with_proofs(
&self,
start_version: u64,
limit: u64,
include_events: bool
) -> Result<Response<Option<TransactionsWithProofsView>>, Error>

pub fn get_account_transactions_with_proofs(
&self,
address: AccountAddress,
start_seq: u64,
limit: u64,
include_events: bool,
ledger_version: Option<u64>
) -> Result<Response<AccountTransactionsWithProofView>, Error>

pub fn get_events_with_proofs(
&self,
key: EventKey,
start_seq: u64,
limit: u64
) -> Result<Response<Vec<EventWithProofView, Global>>, Error>

pub fn get_event_by_version_with_proof(
&self,
key: EventKey,
version: Option<u64>
) -> Result<Response<EventByVersionWithProofView>, Error>

pub fn get_deserialized_events<T>(
&self,
event_key: &EventKey,
start_seq: u64,
limit: u64
) -> Result<Response<Vec<Event<T>, Global>>, Error>where
T: MoveStructType + DeserializeOwned,

Return the events of type T that have been emitted to event_key since start_seq, with a max of limit results Returns an empty vector if there are no such event The type T must match the event types associated with event_key

pub fn get_deserialized_resource<T>(
&self,
address: AccountAddress
) -> Result<Response<Option<T>>, Error>where
T: MoveResource,

Deserialize and return the resource value of type T stored under address Returns None if there is no such value