Struct diem_sdk::client::VerifyingClient

pub struct VerifyingClient<S> { /* fields omitted */ }

The VerifyingClient is a Diem JSON-RPC client that verifies Diem’s cryptographic proofs when it makes API calls.

Concurrency

When issuing multiple concurrent requests with the VerifyingClient, we guarantee:

Each response batch is fulfilled and verified at a ledger version that is greater than or equal to the current trusted version at the time we made the request batch, though not necessarily the globally most recent trusted ledger version.

Requests made serially within a single thread of execution appear strictly ordered, i.e., they were fulfilled and verified at monotonically increasing ledger versions (v1 <= v2 <= ...).

Consequently, without any other effort, multiple concurrent requests may have responses that appear inconsistent or out-of-order. For example, naively making concurrent get_account(..) requests will (most likely) show accounts at different ledger versions; even further, the first response you receive may show a more recent ledger version than the second response.

To avoid this issue, users should pin a concurrent batch of requests to the same ledger version if they want to avoid an inconsistent ledger view.

Implementations

impl<S> VerifyingClient<S> where
S: Storage,

pub fn new(inner: Client, storage: S) -> Result<VerifyingClient<S>, Error>

pub fn new_with_state(
inner: Client,
trusted_state: TrustedState,
storage: S
) -> VerifyingClient<S>

pub fn version(&self) -> u64

Get a snapshot of our current trusted ledger Version.

pub fn waypoint(&self) -> Waypoint

Get a snapshot of our current trusted Waypoint.

pub fn trusted_state(&self) -> TrustedState

Get a snapshot of our current TrustedState.

pub async fn sync(&'_ self) -> Result<(), Error>

Issue get_state_proof requests until we successfully sync to the remote node’s current version (unless we experience a verification error or other I/O error).

pub async fn sync_one_step(&'_ self) -> Result<bool, Error>

Issue a single get_state_proof request and try to verify it. Returns Ok(true) if, after verification, we still need to sync more. Returns Ok(false) if we have finished syncing.

pub fn verify_and_ratchet(
&self,
request_trusted_state: &TrustedState,
state_proof: &(LedgerInfoWithSignatures, EpochChangeProof, AccumulatorConsistencyProof),
maybe_accumulator: Option<&TransactionAccumulatorSummary>
) -> Result<(), Error>

Verify and ratchet forward a request’s trusted state using a state proof and potentially persisting the new trusted state if it is the newest.

pub async fn get_account(
&'_ self,
address: AccountAddress
) -> Result<Response<Option<AccountView>>, Error>

pub async fn get_account_by_version(
&'_ self,
address: AccountAddress,
version: u64
) -> Result<Response<Option<AccountView>>, Error>

pub async fn get_transactions(
&'_ self,
start_version: u64,
limit: u64,
include_events: bool
) -> Result<Response<Vec<TransactionView, Global>>, Error>

pub async fn get_events(
&'_ self,
key: EventKey,
start_seq: u64,
limit: u64
) -> Result<Response<Vec<EventView, Global>>, Error>

pub async fn get_currencies(
&'_ self
) -> Result<Response<Vec<CurrencyInfoView, Global>>, Error>

pub async fn get_network_status(&'_ self) -> Result<Response<u64>, Error>

pub async fn request(
&'_ self,
request: MethodRequest
) -> Result<Response<MethodResponse>, Error>

Send a single request via VerifyingClient::batch.

pub fn actual_batch_size(requests: &[MethodRequest]) -> usize

pub async fn batch(
&'_ self,
requests: Vec<MethodRequest, Global>
) -> Result<Vec<Result<Response<MethodResponse>, Error>, Global>, Error>