Struct diem_sdk::crypto::_once_cell::sync::OnceCell

pub struct OnceCell<T>(_);

A thread-safe cell which can be written to only once.

OnceCell provides & references to the contents without RAII guards.

Reading a non-None value out of OnceCell establishes a happens-before relationship with a corresponding write. For example, if thread A initializes the cell with get_or_init(f), and thread B subsequently reads the result of this call, B also observes all the side effects of f.

Example

use once_cell::sync::OnceCell;

static CELL: OnceCell<String> = OnceCell::new();
assert!(CELL.get().is_none());

std::thread::spawn(|| {
    let value: &String = CELL.get_or_init(|| {
        "Hello, World!".to_string()
    });
    assert_eq!(value, "Hello, World!");
}).join().unwrap();

let value: Option<&String> = CELL.get();
assert!(value.is_some());
assert_eq!(value.unwrap().as_str(), "Hello, World!");

Implementations

impl<T> OnceCell<T>

pub const fn new() -> OnceCell<T>

Creates a new empty cell.

pub fn get(&self) -> Option<&T>

Gets the reference to the underlying value.

Returns None if the cell is empty, or being initialized. This method never blocks.

pub fn get_mut(&mut self) -> Option<&mut T>

Gets the mutable reference to the underlying value.

Returns None if the cell is empty.

pub unsafe fn get_unchecked(&self) -> &T

Get the reference to the underlying value, without checking if the cell is initialized.

Safety

Caller must ensure that the cell is in initialized state, and that the contents are acquired by (synchronized to) this thread.

pub fn set(&self, value: T) -> Result<(), T>

Sets the contents of this cell to value.

Returns Ok(()) if the cell was empty and Err(value) if it was full.

Example

use once_cell::sync::OnceCell;

static CELL: OnceCell<i32> = OnceCell::new();

fn main() {
    assert!(CELL.get().is_none());

    std::thread::spawn(|| {
        assert_eq!(CELL.set(92), Ok(()));
    }).join().unwrap();

    assert_eq!(CELL.set(62), Err(62));
    assert_eq!(CELL.get(), Some(&92));
}

pub fn get_or_init<F>(&self, f: F) -> &T where
F: FnOnce() -> T,

Gets the contents of the cell, initializing it with f if the cell was empty.

Many threads may call get_or_init concurrently with different initializing functions, but it is guaranteed that only one function will be executed.

Panics

If f panics, the panic is propagated to the caller, and the cell remains uninitialized.

It is an error to reentrantly initialize the cell from f. The exact outcome is unspecified. Current implementation deadlocks, but this may be changed to a panic in the future.

Example

use once_cell::sync::OnceCell;

let cell = OnceCell::new();
let value = cell.get_or_init(|| 92);
assert_eq!(value, &92);
let value = cell.get_or_init(|| unreachable!());
assert_eq!(value, &92);

pub fn get_or_try_init<F, E>(&self, f: F) -> Result<&T, E> where
F: FnOnce() -> Result<T, E>,

Gets the contents of the cell, initializing it with f if the cell was empty. If the cell was empty and f failed, an error is returned.

Panics

If f panics, the panic is propagated to the caller, and the cell remains uninitialized.

It is an error to reentrantly initialize the cell from f. The exact outcome is unspecified. Current implementation deadlocks, but this may be changed to a panic in the future.

Example

use once_cell::sync::OnceCell;

let cell = OnceCell::new();
assert_eq!(cell.get_or_try_init(|| Err(())), Err(()));
assert!(cell.get().is_none());
let value = cell.get_or_try_init(|| -> Result<i32, ()> {
    Ok(92)
});
assert_eq!(value, Ok(&92));
assert_eq!(cell.get(), Some(&92))

pub fn take(&mut self) -> Option<T>

Takes the value out of this OnceCell, moving it back to an uninitialized state.

Has no effect and returns None if the OnceCell hasn’t been initialized.

Examples

use once_cell::sync::OnceCell;

let mut cell: OnceCell<String> = OnceCell::new();
assert_eq!(cell.take(), None);

let mut cell = OnceCell::new();
cell.set("hello".to_string()).unwrap();
assert_eq!(cell.take(), Some("hello".to_string()));
assert_eq!(cell.get(), None);

pub fn into_inner(self) -> Option<T>

Consumes the OnceCell, returning the wrapped value. Returns None if the cell was empty.

Examples

use once_cell::sync::OnceCell;

let cell: OnceCell<String> = OnceCell::new();
assert_eq!(cell.into_inner(), None);

let cell = OnceCell::new();
cell.set("hello".to_string()).unwrap();
assert_eq!(cell.into_inner(), Some("hello".to_string()));