Struct diem_sdk::crypto::_once_cell::unsync::Lazy

pub struct Lazy<T, F = fn() -> T> { /* fields omitted */ }

A value which is initialized on the first access.

Example

use once_cell::unsync::Lazy;

let lazy: Lazy<i32> = Lazy::new(|| {
    println!("initializing");
    92
});
println!("ready");
println!("{}", *lazy);
println!("{}", *lazy);

// Prints:
//   ready
//   initializing
//   92
//   92

Implementations

impl<T, F> Lazy<T, F>

pub const fn new(init: F) -> Lazy<T, F>

Creates a new lazy value with the given initializing function.

Example

use once_cell::unsync::Lazy;

let hello = "Hello, World!".to_string();

let lazy = Lazy::new(|| hello.to_uppercase());

assert_eq!(&*lazy, "HELLO, WORLD!");

pub fn into_value(this: Lazy<T, F>) -> Result<T, F>

Consumes this Lazy returning the stored value.

Returns Ok(value) if Lazy is initialized and Err(f) otherwise.

impl<T, F> Lazy<T, F> where
F: FnOnce() -> T,

pub fn force(this: &Lazy<T, F>) -> &T

Forces the evaluation of this lazy value and returns a reference to the result.

This is equivalent to the Deref impl, but is explicit.

Example

use once_cell::unsync::Lazy;

let lazy = Lazy::new(|| 92);

assert_eq!(Lazy::force(&lazy), &92);
assert_eq!(&*lazy, &92);