Crate diem_sdk::crypto::_serde_name

This crate provides a fast and reliable way to extract the Serde name of a Rust container.

#[derive(Deserialize)]
struct Foo {
  bar: Bar,
}

#[derive(Deserialize)]
#[serde(rename = "ABC")]
enum Bar { A, B, C }

assert_eq!(trace_name::<Foo>(), Some("Foo"));
assert_eq!(trace_name::<Bar>(), Some("ABC"));
assert_eq!(trace_name::<Option<Bar>>(), None);

Functions

trace_name
Compute the Serde name of a container.