Trait diem_sdk::crypto::ed25519::ed25519_dalek::Digest

pub trait Digest {
    type OutputSize: ArrayLength<u8>;
    pub fn new() -> Self;

    pub fn update(&mut self, data: impl AsRef<[u8]>);

    pub fn chain(self, data: impl AsRef<[u8]>) -> Self;

    pub fn finalize(self) -> GenericArray<u8, Self::OutputSize>;

    pub fn finalize_reset(&mut self) -> GenericArray<u8, Self::OutputSize>;

    pub fn reset(&mut self);

    pub fn output_size() -> usize;

    pub fn digest(data: &[u8]) -> GenericArray<u8, Self::OutputSize>;
}

The Digest trait specifies an interface common for digest functions.

It’s a convenience wrapper around [Update], [FixedOutput], [Reset], Clone, and Default traits. It also provides additional convenience methods.

Associated Types

type OutputSize: ArrayLength<u8>

Output size for Digest

Loading content...
Required methods

pub fn new() -> Self

Create new hasher instance

pub fn update(&mut self, data: impl AsRef<[u8]>)

Digest data, updating the internal state.

This method can be called repeatedly for use with streaming messages.

pub fn chain(self, data: impl AsRef<[u8]>) -> Self

Digest input data in a chained manner.

pub fn finalize(self) -> GenericArray<u8, Self::OutputSize>

Retrieve result and consume hasher instance.

pub fn finalize_reset(&mut self) -> GenericArray<u8, Self::OutputSize>

Retrieve result and reset hasher instance.

This method sometimes can be more efficient compared to hasher re-creation.

pub fn reset(&mut self)

Reset hasher instance to its initial state.

pub fn output_size() -> usize

Get output size of the hasher

pub fn digest(data: &[u8]) -> GenericArray<u8, Self::OutputSize>

Convenience function to compute hash of the data. It will handle hasher creation, data feeding and finalization.

Example:

println!("{:x}", sha2::Sha256::digest(b"Hello world"));

Loading content...