Trait diem_sdk::crypto::hash::CryptoHasher

pub trait CryptoHasher: Default + Write {
    pub fn seed() -> &'static [u8; 32];

    pub fn update(&mut self, bytes: &[u8]);

    pub fn finish(self) -> HashValue;
}

A trait for representing the state of a cryptographic hasher.

Required methods

pub fn seed() -> &'static [u8; 32]

the seed used to initialize hashing Self before the serialization bytes of the actual value

pub fn update(&mut self, bytes: &[u8])

Write bytes into the hasher.

pub fn finish(self) -> HashValue

Finish constructing the HashValue.

Loading content...