Trait diem_sdk::crypto::ValidCryptoMaterialStringExt

pub trait ValidCryptoMaterialStringExt: ValidCryptoMaterial {
    // Provided methods
    fn from_encoded_string(
        encoded_str: &str
    ) -> Result<Self, CryptoMaterialError> { ... }
    fn to_encoded_string(&self) -> Result<String, Error> { ... }
}

An extension to to/from Strings for ValidCryptoMaterial.

Relies on hex for string encoding / decoding. No required fields, provides a default implementation.

Provided Methods

fn from_encoded_string(encoded_str: &str) -> Result<Self, CryptoMaterialError>

When trying to convert from bytes, we simply decode the string into bytes before checking if we can convert.

fn to_encoded_string(&self) -> Result<String, Error>

A function to encode into hex-string after serializing.