Trait diem_sdk::crypto::traits::PrivateKey

pub trait PrivateKey: Sized {
    type PublicKeyMaterial: PublicKey<PrivateKeyMaterial = Self>;

    fn public_key(&self) -> Self::PublicKeyMaterial { ... }
}

A type family for key material that should remain secret and has an associated type of the PublicKey family.

Required Associated Types

type PublicKeyMaterial: PublicKey<PrivateKeyMaterial = Self>

We require public / private types to be coupled, i.e. their associated type is each other.

Provided Methods

fn public_key(&self) -> Self::PublicKeyMaterial

Returns the associated public key