Trait diem_sdk::crypto::traits::Uniform

pub trait Uniform {
    fn generate<R>(rng: &mut R) -> Self
    where
        R: RngCore + CryptoRng;

    fn generate_for_testing() -> Self
    where
        Self: Sized,
    { ... }
}

A type family for schemes which know how to generate key material from a cryptographically-secure [CryptoRng][::rand::CryptoRng].

Required Methods

fn generate<R>(rng: &mut R) -> Selfwhere
R: RngCore + CryptoRng,

Generate key material from an RNG. This should generally not be used for production purposes even with a good source of randomness. When possible use hardware crypto to generate and store private keys.

Provided Methods

fn generate_for_testing() -> Selfwhere
Self: Sized,

Generate a random key using the shared TEST_SEED