Trait diem_sdk::move_types::gas_schedule::GasAlgebra

pub trait GasAlgebra<GasCarrier> where
    GasCarrier: Add<GasCarrier, Output = GasCarrier> + Sub<GasCarrier, Output = GasCarrier> + Div<GasCarrier, Output = GasCarrier> + Mul<GasCarrier, Output = GasCarrier> + Copy, {
    pub fn new(carrier: GasCarrier) -> Self;

    pub fn get(&self) -> GasCarrier;

    pub fn map<F>(self, f: F) -> Self
    where
        F: Fn(GasCarrier) -> GasCarrier,
    { ... }

    pub fn map2<F>(self, other: impl GasAlgebra<GasCarrier>, f: F) -> Self
    where
        F: Fn(GasCarrier, GasCarrier) -> GasCarrier,
    { ... }

    pub fn app<T, F>(&self, other: &impl GasAlgebra<GasCarrier>, f: F) -> T
    where
        F: Fn(GasCarrier, GasCarrier) -> T,
    { ... }

    pub fn unitary_cast<T>(self) -> T
    where
        T: GasAlgebra<GasCarrier>,
    { ... }

    pub fn add(self, right: impl GasAlgebra<GasCarrier>) -> Self { ... }

    pub fn sub(self, right: impl GasAlgebra<GasCarrier>) -> Self { ... }

    pub fn mul(self, right: impl GasAlgebra<GasCarrier>) -> Self { ... }

    pub fn div(self, right: impl GasAlgebra<GasCarrier>) -> Self { ... }
}

A trait encoding the operations permitted on the underlying carrier for the gas unit, and how other gas-related units can interact with other units – operations can only be performed across units with the same underlying carrier (i.e. as long as the underlying data is the same).

Required methods

pub fn new(carrier: GasCarrier) -> Self

Project a value into the gas algebra.

pub fn get(&self) -> GasCarrier

Get the carrier.

Loading content...
Provided methods

pub fn map<F>(self, f: F) -> Self where
F: Fn(GasCarrier) -> GasCarrier,

Map a function f of one argument over the underlying data.

pub fn map2<F>(self, other: impl GasAlgebra<GasCarrier>, f: F) -> Self where
F: Fn(GasCarrier, GasCarrier) -> GasCarrier,

Map a function f of two arguments over the underlying carrier. Note that this function can take two different implementations of the trait – one for self the other for the second argument. But, we enforce that they have the same underlying carrier.

pub fn app<T, F>(&self, other: &impl GasAlgebra<GasCarrier>, f: F) -> T where
F: Fn(GasCarrier, GasCarrier) -> T,

Apply a function f of two arguments to the carrier. Since f is not an endomorphism, we return the resulting value, as opposed to the result wrapped up in ourselves.

pub fn unitary_cast<T>(self) -> T where
T: GasAlgebra<GasCarrier>,

We allow casting between GasAlgebras as long as they have the same underlying carrier – i.e. they use the same type to store the underlying value.

pub fn add(self, right: impl GasAlgebra<GasCarrier>) -> Self

Add the two GasAlgebras together.

pub fn sub(self, right: impl GasAlgebra<GasCarrier>) -> Self

Subtract one GasAlgebra from the other.

pub fn mul(self, right: impl GasAlgebra<GasCarrier>) -> Self

Multiply two GasAlgebras together.

pub fn div(self, right: impl GasAlgebra<GasCarrier>) -> Self

Divide one GasAlgebra by the other.

Loading content...