Function diem_sdk::transaction_builder::stdlib::encode_rotate_authentication_key_with_nonce_script

pub fn encode_rotate_authentication_key_with_nonce_script(
    sliding_nonce: u64, 
    new_key: Vec<u8, Global>
) -> Script

Rotates the sender’s authentication key to the supplied new authentication key. May be sent by any account that has a sliding nonce resource published under it (usually this is Treasury Compliance or Diem Root accounts).

Technical Description

Rotates the account’s DiemAccount::DiemAccount authentication_key field to new_key. new_key must be a valid ed25519 public key, and account must not have previously delegated its DiemAccount::KeyRotationCapability.

Parameters

Name	Type	Description

account	&signer	Signer reference of the sending account of the transaction.

sliding_nonce	u64	The sliding_nonce (see: SlidingNonce) to be used for this transaction.

new_key	vector<u8>	New ed25519 public key to be used for account.

Common Abort Conditions

Error Category	Error Reason	Description

Errors::NOT_PUBLISHED	SlidingNonce::ESLIDING_NONCE	A SlidingNonce resource is not published under account.

Errors::INVALID_ARGUMENT	SlidingNonce::ENONCE_TOO_OLD	The sliding_nonce is too old and it’s impossible to determine if it’s duplicated or not.

Errors::INVALID_ARGUMENT	SlidingNonce::ENONCE_TOO_NEW	The sliding_nonce is too far in the future.

Errors::INVALID_ARGUMENT	SlidingNonce::ENONCE_ALREADY_RECORDED	The sliding_nonce has been previously recorded.

Errors::INVALID_STATE	DiemAccount::EKEY_ROTATION_CAPABILITY_ALREADY_EXTRACTED	account has already delegated/extracted its DiemAccount::KeyRotationCapability.

Errors::INVALID_ARGUMENT	DiemAccount::EMALFORMED_AUTHENTICATION_KEY	new_key was an invalid length.

Related Scripts

Script::rotate_authentication_key

Script::rotate_authentication_key_with_nonce_admin

Script::rotate_authentication_key_with_recovery_address