Module diem_sdk::types::account_config::constants::diem

Constants

DIEM_MODULE_IDENTIFIER
The Identifier for the Diem module.

Statics

COIN_MODULE

DIEM_MODULE
The ModuleId for the Diem module.

Functions

allowed_currency_code_string
In addition to the constraints for valid Move identifiers, currency codes should consist entirely of uppercase alphanumeric characters (e.g., no underscores).

currency_code_from_type_tag

from_currency_code_string

type_tag_for_currency_code