Trait diem_sdk::crypto::ValidCryptoMaterialStringExt

pub trait ValidCryptoMaterialStringExt: ValidCryptoMaterial {
    pub fn from_encoded_string(
        encoded_str: &str
    ) -> Result<Self, CryptoMaterialError> { ... }

    pub fn to_encoded_string(&self) -> Result<String, Error> { ... }
}

An extension to to/from Strings for ValidCryptoMaterial.

Relies on [hex][::hex] for string encoding / decoding. No required fields, provides a default implementation.

Provided methods

pub fn from_encoded_string(
encoded_str: &str
) -> Result<Self, CryptoMaterialError>

When trying to convert from bytes, we simply decode the string into bytes before checking if we can convert.

pub fn to_encoded_string(&self) -> Result<String, Error>

A function to encode into hex-string after serializing.

Loading content...