Module diem_sdk::types::account_config::constants

Modules

account

addresses

coins

designated_dealer

diem

event

Constants

ACCOUNT_MODULE_IDENTIFIER
The Identifier for the Account module.

CORE_CODE_ADDRESS

DESIGNATED_DEALER_MODULE_IDENTIFIER
The Identifier for the Designated Dealer module.

DIEM_MODULE_IDENTIFIER
The Identifier for the Diem module.

EVENT_HANDLE_GENERATOR_STRUCT_IDENTIFIER

EVENT_HANDLE_STRUCT_IDENTIFIER

EVENT_MODULE_IDENTIFIER

XDX_IDENTIFIER

XDX_NAME

XUS_IDENTIFIER

XUS_NAME

Statics

ACCOUNT_MODULE
The ModuleId for the Account module.

COIN_MODULE

DESIGNATED_DEALER_MODULE
The ModuleId for the Designated Dealer module.

DIEM_MODULE
The ModuleId for the Diem module.

EVENT_MODULE

XDX_MODULE

Functions

allowed_currency_code_string
In addition to the constraints for valid Move identifiers, currency codes should consist entirely of uppercase alphanumeric characters (e.g., no underscores).

coin_name
Return Some(struct_name) if t is a StructTag representing one of the current Diem coin types (XDX, XUS), None otherwise.

currency_code_from_type_tag

diem_root_address

event_handle_generator_struct_tag

from_currency_code_string

reserved_vm_address

testnet_dd_account_address

treasury_compliance_account_address

type_tag_for_currency_code

validator_set_address

xdx_type_tag

xus_tag