Module diem_sdk::types::account_config::resources

Modules

account

balance

chain_id

currency_info

designated_dealer

dual_attestation

freezing_bit

key_rotation_capability

preburn_balance

preburn_queue

preburn_with_metadata

role

role_id

vasp

vasp_domain

withdraw_capability

Structs

AccountResource
A Rust representation of an Account resource. This is not how the Account is represented in the VM but it’s a convenient representation.

BalanceResource
The balance resource held under an account.

ChainIdResource

ChildVASP

Credential

CurrencyInfoResource
Struct that represents a CurrencyInfo resource

DesignatedDealer

FreezingBit

KeyRotationCapabilityResource

Limit
Defines the dual attest limit in microDiem XDX

ParentVASP

PreburnQueueResource
The preburn balance held under an account.

PreburnResource
The preburn balance held under an account.

PreburnWithMetadataResource
The preburn along with metadata held in a preburn queue.

RoleId

VASPDomain

VASPDomainEvent

VASPDomainManager

VASPDomains

WithdrawCapabilityResource

Enums

AccountRole
A enum that captures the collection of role-specific resources stored under each account type

DesignatedDealerPreburns

Constants

VASP_DOMAIN_MODULE_IDENTIFIER
The Identifier for the VASPDomain module.