Module diem_sdk::types::validator_verifier

Structs

ValidatorConsensusInfo

Helper struct to manage validator information for validation

ValidatorVerifier

Supports validation of signatures for known authors with individual voting powers. This struct can be used for all signature verification operations including block and network signature verification, respectively.

Enums

VerifyError

Errors possible during signature verification.

Functions

random_validator_verifier

Helper function to get random validator signers and a corresponding validator verifier for testing. If custom_voting_power_quorum is not None, set a custom voting power quorum amount. With pseudo_random_account_address enabled, logs show 0 -> [0000], 1 -> [1000]